split band into object and handlers
