+++ main.py
@@ -8,7 +8,7 @@ import calview
 import member
 import gig
 import plan
-import band
+import band_handlers
 import help
 import motd
 import credits
@@ -61,9 +61,9 @@ else:
         webapp2.Route('/api/plan/<plan_id:.+>/<plan_attribute:.+>/<new_value:.+>', plan.RestEndpoint),
         webapp2.Route('/api/plan/<plan_id:.+>', plan.RestEndpoint),
 
-        webapp2.Route('/api/bands', band.RestEndpointBands),
-        webapp2.Route('/api/band/members/<band_id:.+>', band.RestEndpointMembers),
-        webapp2.Route('/api/band/<band_id:.+>', band.RestEndpoint),
+        webapp2.Route('/api/bands', band_handlers.RestEndpointBands),
+        webapp2.Route('/api/band/members/<band_id:.+>', band_handlers.RestEndpointMembers),
+        webapp2.Route('/api/band/<band_id:.+>', band_handlers.RestEndpoint),
 
         webapp2.Route('/api/gig/plans/<gig_id:.+>', gig.RestEndpointPlans),
         webapp2.Route('/api/gig/<gig_id:.+>', gig.RestEndpoint),
@@ -76,7 +76,7 @@ else:
         # Standard endpoints
 
         webapp2.Route('/', member.DefaultPage, name='home'),
-        webapp2.Route('/band/<band_name:.+>', band.InfoPage),
+        webapp2.Route('/band/<band_name:.+>', band_handlers.InfoPage),
         webapp2.Route('/login', login.LoginPage, name='login'),
         webapp2.Route('/logout', login.LogoutHandler, name='logout'),
         webapp2.Route('/signup', login.SignupPage, name='signup'),
@@ -144,30 +144,30 @@ else:
         webapp2.Route('/gig_answerlink',gig.AnswerLinkHandler, name="gig_answerlink"),
         webapp2.Route('/print_setlist', gig.PrintSetlist),
         webapp2.Route('/print_planlist', gig.PrintPlanlist),
-        webapp2.Route('/member_spreadsheet', band.MemberSpreadsheet),
-        webapp2.Route('/member_emails', band.MemberEmails),
-        webapp2.Route('/archive_spreadsheet', band.ArchiveSpreadsheet),
+        webapp2.Route('/member_spreadsheet', band_handlers.MemberSpreadsheet),
+        webapp2.Route('/member_emails', band_handlers.MemberEmails),
+        webapp2.Route('/archive_spreadsheet', band_handlers.ArchiveSpreadsheet),
         webapp2.Route('/sendreminder', gig.SendReminder),
-        webapp2.Route('/band_info.html', band.InfoPage),
-        webapp2.Route('/band_edit.html', band.EditPage),
-        webapp2.Route('/band_edit_test_new_member_message', band.TestNewMemberMessage),
-        webapp2.Route('/band_delete.html', band.DeleteBand),
-        webapp2.Route('/band_get_members', band.BandGetMembers, name='getmembers'),
-        webapp2.Route('/band_get_sections', band.BandGetSections, name='getsections'),
-        webapp2.Route('/band_setup_sections', band.SetupSections, name='setupSections'),
-        webapp2.Route('/band_confirm_member', band.ConfirmMember),
-        webapp2.Route('/band_makeadmin', band.AdminMember),
-        webapp2.Route('/band_makeoccasional', band.MakeOccasionalMember),
-        webapp2.Route('/band_removemember', band.RemoveMember),
-        webapp2.Route('/band_admin', band.AdminPage),
-        webapp2.Route('/band_get_member_list', band.GetMemberList),
-        webapp2.Route('/band_nav.html', band.BandNavPage),
-        webapp2.Route('/band_get_upcoming', band.GetUpcoming),
-        webapp2.Route('/band_get_public_members', band.GetPublicMembers),
-        webapp2.Route('/band_invite.html', band.InvitePage),
-        webapp2.Route('/band_send_invites', band.SendInvites),
-        webapp2.Route('/band_gig_archive', band.GigArchivePage),
-        webapp2.Route('/band_gig_trashcan', band.GigTrashcanPage),
+        webapp2.Route('/band_info.html', band_handlers.InfoPage),
+        webapp2.Route('/band_edit.html', band_handlers.EditPage),
+        webapp2.Route('/band_edit_test_new_member_message', band_handlers.TestNewMemberMessage),
+        webapp2.Route('/band_delete.html', band_handlers.DeleteBand),
+        webapp2.Route('/band_get_members', band_handlers.BandGetMembers, name='getmembers'),
+        webapp2.Route('/band_get_sections', band_handlers.BandGetSections, name='getsections'),
+        webapp2.Route('/band_setup_sections', band_handlers.SetupSections, name='setupSections'),
+        webapp2.Route('/band_confirm_member', band_handlers.ConfirmMember),
+        webapp2.Route('/band_makeadmin', band_handlers.AdminMember),
+        webapp2.Route('/band_makeoccasional', band_handlers.MakeOccasionalMember),
+        webapp2.Route('/band_removemember', band_handlers.RemoveMember),
+        webapp2.Route('/band_admin', band_handlers.AdminPage),
+        webapp2.Route('/band_get_member_list', band_handlers.GetMemberList),
+        webapp2.Route('/band_nav.html', band_handlers.BandNavPage),
+        webapp2.Route('/band_get_upcoming', band_handlers.GetUpcoming),
+        webapp2.Route('/band_get_public_members', band_handlers.GetPublicMembers),
+        webapp2.Route('/band_invite.html', band_handlers.InvitePage),
+        webapp2.Route('/band_send_invites', band_handlers.SendInvites),
+        webapp2.Route('/band_gig_archive', band_handlers.GigArchivePage),
+        webapp2.Route('/band_gig_trashcan', band_handlers.GigTrashcanPage),
         webapp2.Route('/calevents', calview.CalEvents),
         webapp2.Route('/updateplan', plan.UpdatePlan),
         webapp2.Route('/updateplanfeedback', plan.UpdatePlanFeedback),
